# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/queue/message": {
      "post": {
        "operationId": "QueueController_addMessage",
        "summary": "Add a message to the queue",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessagePayloadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message successfully added to queue",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "jobId": "12345",
                    "message": "Message added to queue successfully"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid message format"
          }
        },
        "tags": [
          "queue"
        ]
      }
    }
  },
  "info": {
    "title": "NestJS Queue Retry Worker API",
    "description": "API for managing message queues with retry logic and dead letter handling",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "MessagePayloadDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the message",
            "example": "msg_123456"
          },
          "channel": {
            "type": "string",
            "description": "Delivery channel for the message",
            "enum": [
              "http",
              "email",
              "internal"
            ],
            "example": "http"
          },
          "destination": {
            "type": "string",
            "description": "Destination URL, email address, or internal identifier",
            "example": "https://api.example.com/webhook"
          },
          "data": {
            "type": "object",
            "description": "Message payload data",
            "example": {
              "message": "Hello, World!",
              "priority": "high"
            }
          },
          "metadata": {
            "type": "object",
            "description": "Additional metadata for the message",
            "example": {
              "userId": "12345",
              "correlationId": "abc-123"
            }
          }
        },
        "required": [
          "id",
          "channel",
          "destination",
          "data"
        ]
      }
    }
  }
}